renamed module common_function to common_module
added function concurrent_future_get_data
