+++ common_module/work_with_document.py
